+++ client/src/containers/projects/detail/forms/pcb/geographic-scope/index.tsx
@@ -1,0 +1,101 @@
+"use client";
+import { PropsWithChildren } from "react";
+
+import { UseFormReturn, useForm, useWatch } from "react-hook-form";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import parse from "html-react-parser";
+import { ZodTypeAny, z } from "zod";
+
+import { Button } from "@/components/ui/button";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Textarea } from "@/components/ui/textarea";
+
+export interface GeographicScopeFormProps extends PropsWithChildren {
+  items: {
+    data: {
+      id: string;
+      title: string;
+      description: string;
+      display_order: string;
+      category: {
+        display_order: string;
+      };
+    }[];
+  };
+}
+
+const useSyncFormValues = (f: UseFormReturn) => {
+  const values = useWatch({
+    control: f.control,
+  });
+
+  console.info("watch values", values);
+};
+
+export default function GeographicScopeForm({ items }: GeographicScopeFormProps) {
+  const formSchema = z.object({
+    ...items.data.reduce(
+      (acc, { id }) => {
+        acc[id] = z.string().min(10);
+
+        return acc;
+      },
+      {} as Record<string, ZodTypeAny>,
+    ),
+  });
+
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+  });
+
+  const onSubmit = (values: z.infer<typeof formSchema>) => {
+    console.info(values);
+  };
+
+  useSyncFormValues(form);
+
+  return (
+    <Form {...form}>
+      <form
+        onSubmit={form.handleSubmit(onSubmit)}
+        className="space-y-10 border-t border-gray-100 pt-5"
+      >
+        {items.data
+          .sort((a, b) => +a.display_order - +b.display_order)
+          .map(({ id, title, description, display_order, category }) => {
+            return (
+              <FormField
+                key={id}
+                control={form.control}
+                name={`${id}`}
+                render={({ field }) => (
+                  <FormItem className="space-y-2">
+                    <FormLabel>
+                      {`${category.display_order}.${display_order}`} {title}
+                    </FormLabel>
+                    <div className="prose">{parse(description)}</div>
+
+                    <FormControl className="prose mt-5 inline-block border-t border-primary/10 py-2.5">
+                      <Textarea {...field} rows={4} />
+                    </FormControl>
+
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            );
+          })}
+
+        <Button type="submit">Submit</Button>
+      </form>
+    </Form>
+  );
+}
